fix(docker+go): resolve Dockerfile build, gateway, and test failures
- Dockerfile: multi-stage build with proper Go module caching
- docker-compose.yml: fix network collision, port conflicts, service deps
- bos_gateway.go: fix openai/gpt-oss-20b max_tokens for reasoning model
- osasdk/client.go: proper error handling and circuit breaker
- Fix Go test files for Armstrong compliance (no error swallowing)
- otel-collector-config.yaml: proper exporter configuration
- Add weaver semconv service and MCP stdio integration
- Add SemconvWeaverPanel.svelte frontend component

+++ desktop/backend-go/internal/handlers/auth_email_test.go
@@ -180,7 +180,7 @@ func TestLoginMaxAttempts_Constant(t *testing.T) {
 
 func TestLoginLockDuration_Constant(t *testing.T) {
 	// 15 minutes
-	assert.Equal(t, 15*60*1000*1000*1000, loginLockDuration.Nanoseconds())
+	assert.Equal(t, int64(15*60*1000*1000*1000), loginLockDuration.Nanoseconds())
 }
 
 // ---------------------------------------------------------------------------
